BigQuery-GoogleSearch-TopTrends-Dashboard/requirements.txt
pandas
plotly
dash
dash-bootstrap-components
google-cloud-bigquery
flask
